Build 116, ATIK filter wheel updates

+++ src/controller_image.cpp
@@ -107,13 +107,13 @@ ControllerImage::~ControllerImage(void)
 	//*	free up the image memory
 	if (cDownLoadedImage != NULL)
 	{
-		CONSOLE_DEBUG_W_HEX("Release cDownLoadedImage", cDownLoadedImage);
+		CONSOLE_DEBUG_W_HEX("Release cDownLoadedImage", (unsigned long)cDownLoadedImage);
 		cvReleaseImage(&cDownLoadedImage);
 		cDownLoadedImage	=	NULL;
 	}
 	if (cDisplayedImage != NULL)
 	{
-		CONSOLE_DEBUG_W_HEX("Release cDisplayedImage", cDisplayedImage);
+		CONSOLE_DEBUG_W_HEX("Release cDisplayedImage", (unsigned long)cDisplayedImage);
 		cvReleaseImage(&cDisplayedImage);
 		cDisplayedImage	=	NULL;
 	}
